test: split goal gate browser journeys
